@@ -4,6 +4,26 @@ const path = require('path');
 const argv = require('yargs').argv;
 const MANIFEST = 'package.json';
 const exec = require('child_process').exec;
+const through = require('through2');
+const _ = require('lodash');
+const gutil = require('gulp-util');
+
+const MODULE_PATH = './modules';
+const BUILD_PATH = './build/dist';
+const DEV_PATH = './build/dev';
+
+
+// get only subdirectories that contain package.json with 'main' property
+function isModuleDirectory(filePath) {
+  try {
+    const manifestPath = path.join(filePath, MANIFEST);
+    if (fs.statSync(manifestPath).isFile()) {
+      const module = require(manifestPath);
+      return module && module.main;
+    }
+  }
+  catch (error) {}
+}
 
 module.exports = {
   parseBrowserArgs: function (argv) {
@@ -19,6 +39,85 @@ module.exports = {
     return str.replace(/\n/g, '')
         .replace(/<\//g, '<\\/')
         .replace(/\/>/g, '\\/>');
+  },
+  getArgModules() {
+    var modules = (argv.modules || '').split(',').filter(module => !!module);
+
+    try {
+      if (modules.length === 1 && path.extname(modules[0]).toLowerCase() === '.json') {
+        var moduleFile = modules[0];
+
+        modules = JSON.parse(
+          fs.readFileSync(moduleFile, 'utf8')
+        );
+      }
+    } catch(e) {
+      throw new gutil.PluginError({
+        plugin: 'modules',
+        message: 'failed reading: ' + argv.modules
+      });
+    }
+
+    return modules;
+  },
+  getModules: _.memoize(function(externalModules) {
+    externalModules = externalModules || [];
+    var internalModules;
+    try {
+      var absoluteModulePath = path.join(__dirname, MODULE_PATH);
+      internalModules = fs.readdirSync(absoluteModulePath)
+        .filter(file => !(/(^|\/)\.[^\/\.]/g).test(file))
+        .reduce((memo, file) => {
+          var moduleName = file.split(new RegExp('[.\\' + path.sep + ']'))[0];
+          var modulePath = path.join(absoluteModulePath, file);
+          if (fs.lstatSync(modulePath).isDirectory()) {
+            modulePath = path.join(modulePath, "index.js")
+          }
+          memo[modulePath] = moduleName;
+          return memo;
+        }, {});
+    } catch(err) {
+      internalModules = {};
+    }
+    return Object.assign(externalModules.reduce((memo, module) => {
+      try {
+        var modulePath = require.resolve(module);
+        memo[modulePath] = module;
+      } catch(err) {
+        // do something
+      }
+      return memo;
+    }, internalModules));
+  }),
+
+  getBuiltModules: function(dev, externalModules) {
+    var modules = this.getModuleNames(externalModules);
+    if(Array.isArray(externalModules)) {
+      modules = _.intersection(modules, externalModules);
+    }
+    return modules.map(name => path.join(__dirname, dev ? DEV_PATH : BUILD_PATH, name + '.js'));
+  },
+
+  getBuiltPrebidCoreFile: function(dev) {
+    return path.join(__dirname, dev ? DEV_PATH : BUILD_PATH, 'prebid-core' + '.js');
+  },
+
+  getModulePaths: function(externalModules) {
+    var modules = this.getModules(externalModules);
+    return Object.keys(modules);
+  },
+
+  getModuleNames: function(externalModules) {
+    return _.values(this.getModules(externalModules));
+  },
+
+  nameModules: function(externalModules) {
+    var modules = this.getModules(externalModules);
+    return through.obj(function(file, enc, done) {
+      file.named = modules[file.path] ? modules[file.path] : 'prebid';
+      this.push(file);
+      done();
+    })
   },
 
   /*
@@ -37,18 +136,6 @@ module.exports = {
         const module = require(path.join(directory, moduleDirectory, MANIFEST));
         return path.join(directory, moduleDirectory, module.main);
       });
-
-    // get only subdirectories that contain package.json with 'main' property
-    function isModuleDirectory(filePath) {
-      try {
-        const manifestPath = path.join(filePath, MANIFEST);
-        if (fs.statSync(manifestPath).isFile()) {
-          const module = require(manifestPath);
-          return module && module.main;
-        }
-      }
-      catch (error) {}
-    }
   },
 
   createEnd2EndTestReport : function(targetDestinationDir) {
