@@ -101,7 +101,7 @@ module.exports = (function () {
   _public.off = function (eventString, handler, id) {
     var event = _handlers[eventString];
 
-    if (utils.isEmpty(event) || utils.isEmpty(event.que) && utils.isEmpty(event[id])) {
+    if (utils.isEmpty(event) || (utils.isEmpty(event.que) && utils.isEmpty(event[id]))) {
       return;
     }
 
