@@ -1,8 +1,6 @@
-import { getBidRequest, logError, insertPixel } from './utils';
+import { getBidRequest, logError, triggerPixel } from './utils';
 
-/** INSERT NATIVE ADAPTERS - DO NOT EDIT OR REMOVE */
-const nativeAdapters = [];
-/** END INSERT NATIVE ADAPTERS */
+export const nativeAdapters = [];
 
 export const NATIVE_KEYS = {
   title: 'hb_native_title',
@@ -11,6 +9,7 @@ export const NATIVE_KEYS = {
   image: 'hb_native_image',
   icon: 'hb_native_icon',
   clickUrl: 'hb_native_linkurl',
+  cta: 'hb_native_cta',
 };
 
 export const NATIVE_TARGETING_KEYS = Object.keys(NATIVE_KEYS).map(
@@ -35,7 +34,7 @@ const SUPPORTED_TYPES = {
  * passes them on directly. If they were of type 'type', translate
  * them into the predefined specific asset requests for that type of native ad.
  */
-export default function processNativeAdUnitParams(params) {
+export function processNativeAdUnitParams(params) {
   if (params && params.type && typeIsSupported(params.type)) {
     return SUPPORTED_TYPES[params.type];
   }
@@ -79,7 +78,7 @@ export function nativeBidIsValid(bid) {
   const requiredAssets = Object.keys(requestedAssets).filter(
     key => requestedAssets[key].required
   );
-  const returnedAssets = Object.keys(bid.native);
+  const returnedAssets = Object.keys(bid.native).filter(key => bid.native[key]);
 
   return requiredAssets.every(asset => returnedAssets.includes(asset));
 }
@@ -93,6 +92,6 @@ export function fireNativeImpressions(adObject) {
     adObject.native.impressionTrackers;
 
   (impressionTrackers || []).forEach(tracker => {
-    insertPixel(tracker);
+    triggerPixel(tracker);
   });
 }
