@@ -1,8 +1,8 @@
 /** @module polyfill
 Misc polyfills
 */
-/* jshint -W121 */
 require('core-js/fn/array/find');
+require('core-js/fn/array/find-index');
 require('core-js/fn/array/includes');
 require('core-js/fn/object/assign');
 
