@@ -1,19 +1,21 @@
-import { uniques, isGptPubadsDefined, getHighestCpm, adUnitsFilter } from './utils';
+import { uniques, isGptPubadsDefined, getHighestCpm, adUnitsFilter, groupBy } from './utils';
+import { config } from './config';
 import { NATIVE_TARGETING_KEYS } from './native';
-const bidmanager = require('./bidmanager.js');
-const utils = require('./utils.js');
-var CONSTANTS = require('./constants.json');
+const bidmanager = require('./bidmanager');
+const utils = require('./utils');
+var CONSTANTS = require('./constants');
 
 var targeting = exports;
 var pbTargetingKeys = [];
 
 targeting.resetPresetTargeting = function(adUnitCode) {
   if (isGptPubadsDefined()) {
-    const adUnitCodes = adUnitCode && adUnitCode.length ? [adUnitCode] : $$PREBID_GLOBAL$$._adUnitCodes;
+    const adUnitCodes = getAdUnitCodes(adUnitCode);
+    const adUnits = $$PREBID_GLOBAL$$.adUnits.filter(adUnit => adUnitCodes.includes(adUnit.code));
     window.googletag.pubads().getSlots().forEach(slot => {
       pbTargetingKeys.forEach(function(key) {
         // reset only registered adunits
-        adUnitCodes.find(function(unit) {
+        adUnits.forEach(function(unit) {
           if (unit.code === slot.getAdUnitPath() ||
               unit.code === slot.getSlotElementId()) {
             slot.setTargeting(key, null);
@@ -25,13 +27,13 @@ targeting.resetPresetTargeting = function(adUnitCode) {
 };
 
 targeting.getAllTargeting = function(adUnitCode) {
-  const adUnitCodes = adUnitCode && adUnitCode.length ? [adUnitCode] : $$PREBID_GLOBAL$$._adUnitCodes;
+  const adUnitCodes = getAdUnitCodes(adUnitCode);
 
   // Get targeting for the winning bid. Add targeting for any bids that have
   // `alwaysUseBid=true`. If sending all bids is enabled, add targeting for losing bids.
   var targeting = getWinningBidTargeting(adUnitCodes)
-      .concat(getAlwaysUseBidTargeting(adUnitCodes))
-      .concat($$PREBID_GLOBAL$$._sendAllBids ? getBidLandscapeTargeting(adUnitCodes) : []);
+    .concat(getAlwaysUseBidTargeting(adUnitCodes))
+    .concat(config.getConfig('enableSendAllBids') ? getBidLandscapeTargeting(adUnitCodes) : []);
 
   // store a reference of the targeting keys
   targeting.map(adUnitCode => {
@@ -64,9 +66,27 @@ targeting.setTargeting = function(targetingConfig) {
   });
 };
 
+/**
+ * normlizes input to a `adUnit.code` array
+ * @param  {(string|string[])} adUnitCode [description]
+ * @return {string[]}     AdUnit code array
+ */
+function getAdUnitCodes(adUnitCode) {
+  if (typeof adUnitCode === 'string') {
+    return [adUnitCode];
+  } else if (utils.isArray(adUnitCode)) {
+    return adUnitCode;
+  }
+  return $$PREBID_GLOBAL$$._adUnitCodes || [];
+}
+
+/**
+ * Returns top bids for a given adUnit or set of adUnits.
+ * @param  {(string|string[])} adUnitCode adUnitCode or array of adUnitCodes
+ * @return {[type]}            [description]
+ */
 targeting.getWinningBids = function(adUnitCode) {
-  // use the given adUnitCode as a filter if present or all adUnitCodes if not
-  const adUnitCodes = adUnitCode ? [adUnitCode] : $$PREBID_GLOBAL$$._adUnitCodes;
+  const adUnitCodes = getAdUnitCodes(adUnitCode);
 
   return $$PREBID_GLOBAL$$._bidsReceived
     .filter(bid => adUnitCodes.includes(bid.adUnitCode))
@@ -75,13 +95,7 @@ targeting.getWinningBids = function(adUnitCode) {
     .filter(uniques)
     .map(adUnitCode => $$PREBID_GLOBAL$$._bidsReceived
       .filter(bid => bid.adUnitCode === adUnitCode ? bid : null)
-      .reduce(getHighestCpm,
-      {
-        adUnitCode: adUnitCode,
-        cpm: 0,
-        adserverTargeting: {},
-        timeToRespond: 0
-      }));
+      .reduce(getHighestCpm, getEmptyBid(adUnitCode)));
 };
 
 targeting.setTargetingForAst = function() {
@@ -101,8 +115,8 @@ targeting.setTargetingForAst = function() {
   );
 };
 
-function getWinningBidTargeting() {
-  let winners = targeting.getWinningBids();
+function getWinningBidTargeting(adUnitCodes) {
+  let winners = targeting.getWinningBids(adUnitCodes);
   let standardKeys = getStandardKeys();
 
   winners = winners.map(winner => {
@@ -120,9 +134,9 @@ function getWinningBidTargeting() {
 }
 
 function getStandardKeys() {
-  return bidmanager.getStandardBidderAdServerTargeting()        // in case using a custom standard key set
-                   .map(targeting => targeting.key)
-                   .concat(CONSTANTS.TARGETING_KEYS).filter(uniques);     // standard keys defined in the library.
+  return bidmanager.getStandardBidderAdServerTargeting() // in case using a custom standard key set
+    .map(targeting => targeting.key)
+    .concat(CONSTANTS.TARGETING_KEYS).filter(uniques); // standard keys defined in the library.
 }
 
 /**
@@ -152,19 +166,24 @@ function getAlwaysUseBidTargeting(adUnitCodes) {
 
 function getBidLandscapeTargeting(adUnitCodes) {
   const standardKeys = CONSTANTS.TARGETING_KEYS.concat(NATIVE_TARGETING_KEYS);
-
-  return $$PREBID_GLOBAL$$._bidsReceived
-    .filter(adUnitsFilter.bind(this, adUnitCodes))
-    .map(bid => {
-      if (bid.adserverTargeting) {
-        return {
-          [bid.adUnitCode]: getTargetingMap(bid, standardKeys.filter(
-            key => typeof bid.adserverTargeting[key] !== 'undefined') // mainly for possibly
-            // unset hb_deal
-          )
-        };
-      }
-    }).filter(bid => bid); // removes empty elements in array
+  const bids = [];
+  // bucket by adUnitcode
+  let buckets = groupBy($$PREBID_GLOBAL$$._bidsReceived, 'adUnitCode');
+  // filter top bid for each bucket by bidder
+  Object.keys(buckets).forEach(bucketKey => {
+    let bidsByBidder = groupBy(buckets[bucketKey], 'bidderCode');
+    Object.keys(bidsByBidder).forEach(key => bids.push(bidsByBidder[key].reduce(getHighestCpm, getEmptyBid())));
+  });
+  // populate targeting keys for the remaining bids
+  return bids.map(bid => {
+    if (bid.adserverTargeting) {
+      return {
+        [bid.adUnitCode]: getTargetingMap(bid, standardKeys.filter(
+          key => typeof bid.adserverTargeting[key] !== 'undefined')
+        )
+      };
+    }
+  }).filter(bid => bid); // removes empty elements in array
 }
 
 function getTargetingMap(bid, keys) {
@@ -180,3 +199,12 @@ targeting.isApntagDefined = function() {
     return true;
   }
 };
+
+function getEmptyBid(adUnitCode) {
+  return {
+    adUnitCode: adUnitCode,
+    cpm: 0,
+    adserverTargeting: {},
+    timeToRespond: 0
+  };
+}
