@@ -1106,7 +1106,7 @@ export function getBidResponsesFromAPI() {
   };
 }
 
-// Ad server targeting when `$$PREBID_GLOBAL$$.enableSendAllBids()` is called.
+// Ad server targeting when `setConfig({ enableSendAllBids: true })` is set.
 export function getAdServerTargeting() {
   return {
     '/19968336/header-bid-tag-0': {
@@ -1368,5 +1368,15 @@ export function getBidRequestedPayload() {
       }
     ],
     'start': 1465426155412
+  };
+}
+
+export function getCurrencyRates() {
+  return {
+    'dataAsOf': '2017-04-25',
+    'conversions': {
+      'GBP': { 'CNY': 8.8282, 'JPY': 141.7, 'USD': 1.2824 },
+      'USD': { 'CNY': 6.8842, 'GBP': 0.7798, 'JPY': 110.49 }
+    }
   };
 }
