@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import {getPriceBucketString, isValidePriceConfig} from 'src/cpmBucketManager';
+import {getPriceBucketString, isValidPriceConfig} from 'src/cpmBucketManager';
 let cpmFixtures = require('test/fixtures/cpmInputsOutputs.json');
 
 describe('cpmBucketManager', () => {
@@ -35,6 +35,62 @@ describe('cpmBucketManager', () => {
     expect(JSON.stringify(output)).to.deep.equal(expected);
   });
 
+  it('gets the correct custom bucket strings in non-USD currency', () => {
+    let cpm = 16.50908 * 110.49;
+    let customConfig = {
+      'buckets': [{
+        'precision': 4,
+        'min': 0,
+        'max': 3,
+        'increment': 0.01,
+      },
+      {
+        'precision': 4,
+        'min': 3,
+        'max': 18,
+        'increment': 0.05,
+        'cap': true
+      }
+      ]
+    };
+    let expected = '{"low":"552.45","med":"1823.09","high":"1823.09","auto":"1823.09","dense":"1823.09","custom":"1823.0850"}';
+    let output = getPriceBucketString(cpm, customConfig, 110.49);
+    expect(JSON.stringify(output)).to.deep.equal(expected);
+  });
+
+  it('gets custom bucket strings and it should honor 0', () => {
+    let cpm = 16.50908;
+    let customConfig = {
+      'buckets': [
+        {
+          'precision': 0,
+          'min': 3,
+          'max': 18,
+          'increment': 0.05,
+        }
+      ]
+    };
+    let expected = '{"low":"5.00","med":"16.50","high":"16.50","auto":"16.50","dense":"16.50","custom":"17"}';
+    let output = getPriceBucketString(cpm, customConfig);
+    expect(JSON.stringify(output)).to.deep.equal(expected);
+  });
+
+  it('gets the custom bucket strings without passing precision and it should honor the default precision', () => {
+    let cpm = 16.50908;
+    let customConfig = {
+      'buckets': [
+        {
+          'min': 3,
+          'max': 18,
+          'increment': 0.05,
+        }
+      ]
+    };
+    let expected = '{"low":"5.00","med":"16.50","high":"16.50","auto":"16.50","dense":"16.50","custom":"16.50"}';
+    let output = getPriceBucketString(cpm, customConfig);
+    expect(JSON.stringify(output)).to.deep.equal(expected);
+  });
+
   it('checks whether custom config is valid', () => {
     let badConfig = {
       'buckets': [{
@@ -43,7 +99,7 @@ describe('cpmBucketManager', () => {
         'increment': 0.01,
       },
       {
-          // missing min prop
+        // missing min prop
         'max': 18,
         'increment': 0.05,
         'cap': true
@@ -51,6 +107,6 @@ describe('cpmBucketManager', () => {
       ]
     };
 
-    expect(isValidePriceConfig(badConfig)).to.be.false;
+    expect(isValidPriceConfig(badConfig)).to.be.false;
   });
 });
