@@ -6,7 +6,7 @@ describe('helpers.url', () => {
     let parsed;
 
     beforeEach(() => {
-      parsed = parse('http://example.com:3000/pathname/?search=test&foo=bar#hash');
+      parsed = parse('http://example.com:3000/pathname/?search=test&foo=bar&bar=foo%26foo%3Dxxx#hash');
     });
 
     it('extracts the protocol', () => {
@@ -28,8 +28,9 @@ describe('helpers.url', () => {
     it('extracts the search query', () => {
       expect(parsed).to.have.property('search');
       expect(parsed.search).to.eql({
-        foo: 'bar',
-        search: 'test'
+        foo: 'xxx',
+        search: 'test',
+        bar: 'foo',
       });
     });
 
@@ -42,6 +43,23 @@ describe('helpers.url', () => {
     });
   });
 
+  describe('parse(url, {noDecodeWholeURL: true})', () => {
+    let parsed;
+
+    beforeEach(() => {
+      parsed = parse('http://example.com:3000/pathname/?search=test&foo=bar&bar=foo%26foo%3Dxxx#hash', {noDecodeWholeURL: true});
+    });
+
+    it('extracts the search query', () => {
+      expect(parsed).to.have.property('search');
+      expect(parsed.search).to.eql({
+        foo: 'bar',
+        search: 'test',
+        bar: 'foo%26foo%3Dxxx',
+      });
+    });
+  });
+
   describe('format()', () => {
     it('formats an object in to a URL', () => {
       expect(format({
@@ -49,9 +67,9 @@ describe('helpers.url', () => {
         hostname: 'example.com',
         port: 3000,
         pathname: '/pathname/',
-        search: {foo: 'bar', search: 'test'},
+        search: {foo: 'bar', search: 'test', bar: 'foo%26foo%3Dxxx'},
         hash: 'hash'
-      })).to.equal('http://example.com:3000/pathname/?foo=bar&search=test#hash');
+      })).to.equal('http://example.com:3000/pathname/?foo=bar&search=test&bar=foo%26foo%3Dxxx#hash');
     });
 
     it('will use defaults for missing properties', () => {
